Completed work to move to server-side processing and data storage.
app.js

* Now queries the server's state on startup.

components/fitness.js

* Now pulls genome/features from server when selected from dropdowns.
* Has loading screen

components/help.js

* Updated to be slightly more clear about import/export functionality.

components/ingest.js

* Now defers to server for data that's available. Proxies commands about
ingestion or removal to the server.
* Import/export is now based on creating a file on the user's computer.
This is to avoid having to have the browser hold and serialize the state
in memoroy, which was the point of this rewrite in the first place.

components/visualizer.js

* now downloads genome from server when selected from dropdown
* now downloads measurements from server when "Add" is clicked

lib/analysis.js

* communication with the server for ingesting data is more coherent.

reducers/ingest.js

* drastically simplified by moving logic to server. Now we just update
the entire state.